Add "Join our Newsletter" link to footer
This commit also adds a Gemfile for easier development, in service of
adding a simple "Join our Newsletter" link to every page's footer.

I decided to replace the "write an email" option here because the "write
us an email" CTA is already in the "get involved" partial and right
above these links on most pages anyway.